+++ CHANGELOG.md
@@ -8,4 +8,6 @@
 
 - [command_toolbox.warp] feature: new `/warp set-position` command, to change the `position` of a warp conveniently.
 - [command_advice] feature: add `document` property, to make the config clear.
+- [core] new features
+  - feature: new `hotkey F` for the `search button` in all paged GUI.